+++ CHANGELOG.md
@@ -49,6 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a custom block that does LayerNorm scaling.
+- Added `OLMo-mix-0625-150Bsample` data mix.
+- Added alias support to `DataMix` enum.
 - Added the `HalfCos` learning rate scheduler.
 - Added `CONTRIBUTING.md` guidelines.
 - Added a lightweight, gantry-like Beaker launch CLI: `python -m olmo_core.launch.beaker`.
@@ -63,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `olmo3_7B` model config.
 - Added additional internal configuration tools.
 - Added a new named data mix that we used for the 32B run
+- Added internal OLMo3 7B midtraining config.
 
 
 ## [v2.2.0](https://github.com/allenai/OLMo-core/releases/tag/v2.2.0) - 2025-08-26